Working for dotnet core, now how the hell do you make a cross compatible package
