+++ TS/TS.CodeGenerator.Console/Program.cs
@@ -1,0 +1,41 @@
+﻿using System;
+using System.IO;
+
+namespace TS.CodeGenerator.Console
+{
+    class Program
+    {
+        static void Main(string[] args)
+        {
+            var input = Path.GetFullPath(args[0]);
+            var output = Path.GetFullPath(args[1]);
+
+            System.Console.WriteLine($"Input Path {input}");
+            System.Console.WriteLine($"Output Path {output}");
+
+            if (!File.Exists(input))
+            {
+                System.Console.Error.WriteLine($"Could Not Find input {input}");
+                return;
+            }
+           
+            Settings.MethodReturnTypeFormatString = "{0}";
+            var reader = new AssemblyReader(input);
+
+            if (File.Exists(output))
+            {
+                File.Delete(output);
+            }
+            using (var of = File.OpenWrite(output))
+            using (var sw = new StreamWriter(of))
+            {
+                var types = reader.GenerateTypingsString();
+               // sw.WriteLine(@"/// <reference path=""../jquery/jquery.d.ts"" />");
+                sw.WriteLine(types);
+                
+            }
+            System.Console.WriteLine("...");
+            System.Console.WriteLine("Completed");
+        }
+    }
+}
