+++ TS/TS.CodeGenerator.Console/TS.CodeGenerator.Console.csproj
@@ -1,0 +1,13 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <OutputType>Exe</OutputType>
+    <TargetFramework>netcoreapp1.1</TargetFramework>
+    
+  </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\TS.CodeGenerator.Core\TS.CodeGenerator.Core.csproj" />
+  </ItemGroup>
+
+</Project>
