+++ TS/TS.CodeGenerator.Core/IAssemblyReader.cs
