+++ TS/TS.CodeGenerator.Core/IGenerateTS.cs
