+++ TS/TS.CodeGenerator.Core/ITSGenerator.cs
