+++ TS/TS.CodeGenerator.Core/Settings.cs
