+++ TS/TS.CodeGenerator.Tests.Core/TSEnumeration_tests.cs
