+++ TS/TS.CodeGenerator.Tests.Core/TSInterfact_test.cs
