--- TS/lib.test.a.b/packages.config
@@ -1,4 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="TS.CodeGenerator" version="1.0.0.14" targetFramework="net40" />
-</packages>
