--- TS/test.web/Web.config
@@ -1,14 +1,0 @@
-﻿<?xml version="1.0"?>
-
-<!--
-  For more information on how to configure your ASP.NET application, please visit
-  http://go.microsoft.com/fwlink/?LinkId=169433
-  -->
-
-<configuration>
-    <system.web>
-      <compilation debug="true" targetFramework="4.5" />
-      <httpRuntime targetFramework="4.5" />
-    </system.web>
-
-</configuration>
