--- TS/test.web/packages.config
@@ -1,4 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="jquery.TypeScript.DefinitelyTyped" version="2.2.6" targetFramework="net45" />
-</packages>
