--- TS/test.web/test.js
@@ -1,8 +1,0 @@
-var x = (function () {
-    function x() {
-        this.data = [];
-        this.bob = 3;
-    }
-    return x;
-})();
-//# sourceMappingURL=test.js.map
