--- TS/test.web/test.ts
@@ -1,5 +1,0 @@
-﻿ 
- class x implements lib.test.a.b.ITestGenericClass<string> {
-     data: number[] = [];
-     bob=3;
- }
